requirements.txt
# FastAPI
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6

# Database
sqlalchemy>=2.0.25
alembic>=1.13.1
psycopg2-binary>=2.9.9

# Auth
firebase-admin>=6.4.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4

# RAG
weaviate-client>=4.9.0
groq>=0.11.0

# Web scraping
beautifulsoup4>=4.12.3
trafilatura>=1.6.4
requests>=2.31.0

# Utilities
python-dotenv>=1.0.0
pydantic-settings>=2.1.0